pyproject.toml
[tool.poetry]
name = "word2lip"
version = "0.1.0"
description = ""
authors = ["Oleg <pyataev@ieg.net.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^7.2.2"
click = "^8.1.3"
opencv-python = "^4.7.0"
pandas = "^1.5.3"
Pillow = "^9.3.0"
ImageHash = "^4.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
